Update README.md
Hey this document is great. I'd like to propose this change to make it more clear. As it is currently written there is a single phrase which I believe is too broad in scope and would limit legitimate and positive conversations which are within the scope of this community.

>Examples of unacceptable behavior by participants include:
> ...
>Verbal or written comments related to gender, sexual orientation, disability, physical appearance, body size, race or religion

As I interpret this phrase, any discussion around the topics listed above is unacceptable. This would limit, for example,  discussions of gender around pay and equal treatment in the workplace, or discussions about accessibility or the needs of users with disabilities.

If this is clarified that the comments are negative in nature, or directed to an individual I believe it may be more helpful. I'm not sure if my verbiage is perfect, but maybe a good starting place?